hotfix: int, uint => int256, uin256 for consistency

+++ contracts/PriceManager.sol
@@ -17,11 +17,11 @@ contract PriceManager is IPriceManager {
     L3Vault public l3Vault;
 
     mapping(address => bool) public isKeeper;
-    mapping(uint => uint) public indexPrice;
-    mapping(uint => uint) public priceBufferUpdatedTime;
-    mapping(uint => int) public lastPriceBuffer;
+    mapping(uint256 => uint256) public indexPrice;
+    mapping(uint256 => uint256) public priceBufferUpdatedTime;
+    mapping(uint256 => int256) public lastPriceBuffer;
 
-    event Execution(uint assetId, int price);
+    event Execution(uint256 assetId, int256 price);
 
     constructor(address _keeperAddress, address _l3VaultAddress) {
         isKeeper[_keeperAddress] = true;
@@ -39,11 +39,11 @@ contract PriceManager is IPriceManager {
     ) external override onlyKeeper {
         require(_assetId.length == _price.length, "Wrong input");
         uint256 l = _assetId.length;
-        for (uint i = 0; i < uint(l); i++) {
+        for (uint256 i = 0; i < uint256(l); i++) {
             require(_price[i] > 0, "price has to be positive");
 
-            int currentPriceBuffer = getPriceBuffer(_assetId[i]); // % of price shift
-            int currentPriceBufferInUsd = (int256(_price[i]) *
+            int256 currentPriceBuffer = getPriceBuffer(_assetId[i]); // % of price shift
+            int256 currentPriceBufferInUsd = (int256(_price[i]) *
                 currentPriceBuffer) / int256(PRICE_BUFFER_PRECISION);
 
             // int prevMarkPrice = indexPrice[_assetId[i]] +
@@ -65,20 +65,23 @@ contract PriceManager is IPriceManager {
                 );
 
             // TODO: set price with markPriceWithLimitOrderPriceImpact
-            int newPriceBuffer = ((int256(markPriceWithLimitOrderPriceImpact) -
-                int256(_price[i])) * int256(PRICE_BUFFER_PRECISION)) /
+            int256 newPriceBuffer = ((int256(
+                markPriceWithLimitOrderPriceImpact
+            ) - int256(_price[i])) * int256(PRICE_BUFFER_PRECISION)) /
                 int256(_price[i]);
             setPriceBuffer(_assetId[i], newPriceBuffer);
             indexPrice[_assetId[i]] = _price[i];
         }
     }
 
-    function getPriceBuffer(uint _assetId) public view override returns (int) {
-        int elapsedTime = int(
+    function getPriceBuffer(
+        uint256 _assetId
+    ) public view override returns (int256) {
+        int256 elapsedTime = int256(
             block.timestamp - priceBufferUpdatedTime[_assetId]
         );
-        int decayedAmount = elapsedTime * int256(DECAY_CONSTANT);
-        int absLastPriceBuffer = lastPriceBuffer[_assetId] >= 0
+        int256 decayedAmount = elapsedTime * int256(DECAY_CONSTANT);
+        int256 absLastPriceBuffer = lastPriceBuffer[_assetId] >= 0
             ? lastPriceBuffer[_assetId]
             : -lastPriceBuffer[_assetId];
         if (decayedAmount >= absLastPriceBuffer) {
@@ -91,42 +94,42 @@ contract PriceManager is IPriceManager {
         }
     }
 
-    function setPriceBuffer(uint _assetId, int value) internal {
+    function setPriceBuffer(uint256 _assetId, int256 value) internal {
         lastPriceBuffer[_assetId] = value;
         priceBufferUpdatedTime[_assetId] = block.timestamp;
     }
 
     function getIndexPrice(
-        uint _assetId
+        uint256 _assetId
     ) external view override returns (uint256) {
         return indexPrice[_assetId];
     }
 
     function getMarkPrice(
-        uint _assetId
+        uint256 _assetId
     ) public view override returns (uint256) {
-        int newPriceBuffer = getPriceBuffer(_assetId);
-        int newPriceBufferInUsd = (int256(indexPrice[_assetId]) *
+        int256 newPriceBuffer = getPriceBuffer(_assetId);
+        int256 newPriceBufferInUsd = (int256(indexPrice[_assetId]) *
             newPriceBuffer) / int256(PRICE_BUFFER_PRECISION);
         return uint256(int256(indexPrice[_assetId]) + newPriceBufferInUsd);
     }
 
     function getAverageExecutionPrice(
-        uint _assetId,
-        uint _size,
+        uint256 _assetId,
+        uint256 _size,
         bool _isBuy
     ) external override returns (uint256) {
-        uint price = getMarkPrice(_assetId);
+        uint256 price = getMarkPrice(_assetId);
         // require first bit of _size is 0
         require(_size < 2 ** 255, "size overflow");
         require(price > 0, "price not set");
-        int intSize = _isBuy ? int(_size) : -int(_size);
-        int priceBufferChange = intSize / int256(PRICE_BUFFER_DELTA_TO_SIZE);
+        int256 intSize = _isBuy ? int256(_size) : -int256(_size);
+        int256 priceBufferChange = intSize / int256(PRICE_BUFFER_DELTA_TO_SIZE);
         setPriceBuffer(_assetId, getPriceBuffer(_assetId) + priceBufferChange);
-        int averageExecutedPrice = int(price) +
-            (int(price) * priceBufferChange) /
+        int256 averageExecutedPrice = int256(price) +
+            (int256(price) * priceBufferChange) /
             2 /
-            int(PRICE_BUFFER_PRECISION);
+            int256(PRICE_BUFFER_PRECISION);
         require(averageExecutedPrice > 0, "price underflow");
         emit Execution(_assetId, averageExecutedPrice);
         return uint256(averageExecutedPrice);
